Hook useGroupedEvents
